Extract ASM80 acceptance test helpers
