+++ test/asm80/mon3_acceptance.test.ts
@@ -1,15 +1,20 @@
-import { copyFileSync, existsSync, mkdtempSync, readFileSync, readdirSync, rmSync } from 'node:fs';
-import { basename, dirname, join, resolve } from 'node:path';
-import { spawnSync } from 'node:child_process';
-import { tmpdir } from 'node:os';
+import { existsSync } from 'node:fs';
+import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import { describe, expect, it } from 'vitest';
 
 import { compile } from '../../src/compile.js';
-import type { Diagnostic } from '../../src/diagnosticTypes.js';
 import { defaultFormatWriters } from '../../src/formats/index.js';
 import type { BinArtifact } from '../../src/formats/types.js';
+import {
+  copyZ80Siblings,
+  findAsm80Executable,
+  findFirstMismatch,
+  runAsm80Reference,
+  summarizeBinaryMismatch,
+  summarizeDiagnostics,
+} from '../helpers/index.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -21,24 +26,7 @@ const manifest = {
   source: process.env.MON3_SOURCE ?? '/Users/johnhardy/Documents/projects/MON3/src/mon3.z80',
 };
 
-function normalizeExecutableCandidate(candidate: string): string {
-  return candidate.includes('/') || candidate.includes('\\') ? resolve(candidate) : candidate;
-}
-
-const asm80Candidates = [
-  process.env.ASM80,
-  process.env.ASM80_PATH,
-  '/Users/johnhardy/Documents/projects/debug80/node_modules/.bin/asm80',
-  'asm80',
-]
-  .filter(
-    (candidate): candidate is string => candidate !== undefined && candidate.trim().length > 0,
-  )
-  .map(normalizeExecutableCandidate);
-const asm80 = asm80Candidates.find((candidate) => {
-  const probe = spawnSync(candidate, ['-h'], { encoding: 'utf8' });
-  return !probe.error;
-});
+const asm80 = findAsm80Executable();
 const mon3FilesAvailable = existsSync(manifest.source);
 const runMon3Acceptance = process.env.ZAX_RUN_MON3_ACCEPTANCE === '1';
 const describeMon3 =
@@ -50,88 +38,14 @@ const describeMon3 =
     ? describe
     : describe.skip;
 
-function byteHex(value: number | undefined): string {
-  return value === undefined ? 'EOF' : `0x${value.toString(16).padStart(2, '0')}`;
-}
-
-function offsetHex(offset: number): string {
-  return `0x${offset.toString(16).padStart(4, '0')}`;
-}
-
-function diagnosticLocation(diagnostic: Diagnostic): string {
-  if (diagnostic.line === undefined || diagnostic.column === undefined) return diagnostic.file;
-  return `${diagnostic.file}:${diagnostic.line}:${diagnostic.column}`;
-}
-
-function summarizeDiagnostics(diagnostics: Diagnostic[], limit = 3): string {
-  const preview = diagnostics
-    .slice(0, limit)
-    .map(
-      (diagnostic) =>
-        `${diagnosticLocation(diagnostic)}: ${diagnostic.severity} [${diagnostic.id}] ${
-          diagnostic.message
-        }`,
-    );
-  return [
-    `Diagnostics preview (showing ${preview.length} of ${diagnostics.length}):`,
-    ...preview,
-  ].join('\n');
-}
-
-function findFirstMismatch(actual: Buffer, reference: Buffer): number {
-  const maxLength = Math.max(actual.length, reference.length);
-  for (let i = 0; i < maxLength; i++) {
-    if (actual[i] !== reference[i]) return i;
-  }
-  return -1;
-}
-
-function summarizeBinaryMismatch(actual: Buffer, reference: Buffer): string {
-  const firstMismatch = findFirstMismatch(actual, reference);
-  const lines = [`Binary length: actual=${actual.length} reference=${reference.length}`];
-  if (firstMismatch >= 0) {
-    lines.push(
-      `First mismatch @${offsetHex(firstMismatch)}: actual=${byteHex(
-        actual[firstMismatch],
-      )} reference=${byteHex(reference[firstMismatch])}`,
-    );
-  } else {
-    lines.push('First mismatch: none');
-  }
-  return lines.join('\n');
-}
-
-function copyAsm80SourceTree(source: string, outDir: string): void {
-  for (const entry of readdirSync(dirname(source))) {
-    if (entry.toLowerCase().endsWith('.z80')) {
-      copyFileSync(join(dirname(source), entry), join(outDir, entry));
-    }
-  }
-}
-
 function buildAsm80Reference(source: string): Buffer {
-  if (!asm80) throw new Error('asm80 executable not found');
-  const outDir = mkdtempSync(join(tmpdir(), 'zax-mon3-asm80-reference-'));
-  const outName = 'mon3-reference.bin';
-  const outBin = join(outDir, outName);
-  try {
-    copyAsm80SourceTree(source, outDir);
-    const result = spawnSync(asm80, ['-m', 'Z80', '-t', 'bin', '-o', outName, basename(source)], {
-      cwd: outDir,
-      encoding: 'utf8',
-    });
-    if (result.error) throw result.error;
-    if (result.status !== 0) {
-      throw new Error(
-        [`asm80 failed with status ${result.status}`, result.stdout.trim(), result.stderr.trim()]
-          .filter((part) => part.length > 0)
-          .join('\n'),
-      );
-    }
-    return readFileSync(outBin);
-  } finally {
-    rmSync(outDir, { recursive: true, force: true });
-  }
+  return runAsm80Reference({
+    asm80,
+    source,
+    tempPrefix: 'zax-mon3-asm80-reference-',
+    outputName: 'mon3-reference.bin',
+    prepareSourceTree: copyZ80Siblings,
+  });
 }
 
 describe('MON3 acceptance failure summaries', () => {
