+++ test/helpers/index.ts
@@ -1,5 +1,6 @@
 export * from './diagnostics.js';
 export * from './cli.js';
 export * from './cliBuild.js';
+export * from './asm80_acceptance.js';
 export * from './lowered_program.js';
 export * from './setup.js';
